Fix path to bundle.js
The path to the compiled Javascript file was wrong: it tries to go up a directory when it shouldn't. Amazingly the current version still works if you host directly from the `dist` folder I guess because there is no dir to go up to?

Anyway, this should make it work in all cases.